// frontend/StockManager/generated-sources/openapi/models/StockDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * StockManager API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface StockDto
 */
export interface StockDto {
    /**
     * 
     * @type {number}
     * @memberof StockDto
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof StockDto
     */
    symbol?: string | null;
    /**
     * 
     * @type {string}
     * @memberof StockDto
     */
    companyName?: string | null;
    /**
     * 
     * @type {string}
     * @memberof StockDto
     */
    sector?: string | null;
    /**
     * 
     * @type {number}
     * @memberof StockDto
     */
    price?: number | null;
    /**
     * 
     * @type {number}
     * @memberof StockDto
     */
    marketCap?: number;
    /**
     * 
     * @type {number}
     * @memberof StockDto
     */
    stockDataId?: number;
}

/**
 * Check if a given object implements the StockDto interface.
 */
export function instanceOfStockDto(value: object): value is StockDto {
    return true;
}

export function StockDtoFromJSON(json: any): StockDto {
    return StockDtoFromJSONTyped(json, false);
}

export function StockDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): StockDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'symbol': json['symbol'] == null ? undefined : json['symbol'],
        'companyName': json['companyName'] == null ? undefined : json['companyName'],
        'sector': json['sector'] == null ? undefined : json['sector'],
        'price': json['price'] == null ? undefined : json['price'],
        'marketCap': json['marketCap'] == null ? undefined : json['marketCap'],
        'stockDataId': json['stockDataId'] == null ? undefined : json['stockDataId'],
    };
}

export function StockDtoToJSON(json: any): StockDto {
    return StockDtoToJSONTyped(json, false);
}

export function StockDtoToJSONTyped(value?: StockDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'symbol': value['symbol'],
        'companyName': value['companyName'],
        'sector': value['sector'],
        'price': value['price'],
        'marketCap': value['marketCap'],
        'stockDataId': value['stockDataId'],
    };
}